neoserve_ai/requirements.txt
# Core dependencies
fastapi>=0.95.2,<1.0.0
uvicorn>=0.22.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=1.10.7,<2.0.0
python-multipart>=0.0.6,<1.0.0

# Google Cloud dependencies
google-cloud-aiplatform>=1.28.1,<2.0.0
google-cloud-bigquery>=3.11.4,<4.0.0
google-cloud-firestore>=2.11.1,<3.0.0
google-cloud-logging>=3.5.0,<4.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0

# Async
anyio>=3.6.2,<4.0.0
httpx>=0.24.1,<1.0.0

# Data processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
